# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: KatalogitAI API
  version: 1.0.0
  description: >
    Backend for KatalogitAI — AI product photoshoots for fashion vendors.
    All endpoints (except webhooks) require `Authorization: Bearer <Firebase ID token>`.
    The credit ledger lives in Cloud SQL; catalogue lives in Firestore.
servers:
  - url: /api/v1
    description: Same-origin via Firebase Hosting rewrite to Cloud Run
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        data: { type: object }
      required: [ok]
    ApiError:
      type: object
      properties:
        ok: { type: boolean, enum: [false] }
        error:
          type: object
          properties:
            code: { type: string, example: INSUFFICIENT_CREDITS }
            message: { type: string }
...
    Job:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [model_shoot, regen_shot, full_regen, retake] }
        productId: { type: string }
        status: { type: string, enum: [queued, processing, done, failed, refunded] }
        creditsReserved: { type: number }
        outputPaths: { type: array, items: { type: string } }
        errorCode: { type: string, nullable: true }
    JobCreateRequest:
      type: object
      required: [type, productId]
      properties:
        type: { type: string, enum: [model_shoot, regen_shot, full_regen, retake] }
        productId: { type: string }
...
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data: { type: object }
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
security:
  - bearerAuth: []
paths:
...
  /jobs:
    post:
      summary: Create an AI job (reserves credits atomically)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCreateRequest' }
      responses:
        '202':
          description: Job queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Job' }
        '402': { $ref: '#/components/responses/Error' }
  /jobs/{jobId}:
    get:
      summary: Poll job status
      parameters: [{ $in: path, name: jobId, required: true, schema: { type: string } }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...